refactor(canvas): consume chat target domain directly
