+++ apps/canvas-workspace/src/main/__tests__/bundle-boundaries.test.ts
@@ -58,8 +58,8 @@ const DYNAMIC_ONLY_MODULE_SUFFIXES = [
   '/components/node-bodies/IframeNodeBody/index.tsx',
   '/components/chat/ChatPanel.tsx',
   '/components/chat/chatScope.ts',
-  '/components/chat/useRegisterChatTarget.ts',
-  '/components/chat/utils/sessionScope.ts',
+  '/agent-chat/target/useRegisterChatTarget.ts',
+  '/agent-chat/target/sessionScope.ts',
   '/components/dock/ReferenceDrawer/index.tsx',
   '/components/canvas/CommandPalette/index.tsx',
   '/components/canvas/SearchBar/index.tsx',
