+++ apps/canvas-workspace/src/renderer/src/components/canvas/Canvas/CanvasSurface.tsx
@@ -22,7 +22,7 @@ import type { CanvasNodeRenderMode } from '../CanvasNodeView/types';
 import type { MindmapTransferHandlers } from '../../../utils/mindmapTransfer';
 import { markOnce } from '../../../perf/monitor';
 import { CanvasGestureHud, MarqueePreview, ShapeDraftPreview } from './CanvasGestureOverlays';
-import type { ChatDeliveryReceipt } from '../../chat/ChatTargetContext';
+import type { ChatDeliveryReceipt } from '../../../agent-chat/target';
 
 const FIT_TRANSITION =
   'transform 0.32s cubic-bezier(0.25, 0.46, 0.45, 0.94), --canvas-scale 0.32s cubic-bezier(0.25, 0.46, 0.45, 0.94)';
