+++ apps/canvas-workspace/src/renderer/src/components/canvas/CanvasNodeView/CanvasNodeHeader.tsx
@@ -23,7 +23,7 @@ import {
 import { NodeTypeBadge } from './NodeTypeBadge';
 import { isKnowledgeNodeType } from '../../../views/WorkspaceNodes/utils';
 import { isReferenceableNode } from '../../../utils/referenceNodes';
-import type { ChatDeliveryReceipt } from '../../chat/ChatTargetContext';
+import type { ChatDeliveryReceipt } from '../../../agent-chat/target';
 
 interface CanvasNodeHeaderProps {
   fullscreenButton: ReactNode;
