+++ apps/canvas-workspace/src/renderer/src/components/canvas/CanvasNodeView/useCanvasNodeViewModel.test.tsx
@@ -5,7 +5,7 @@ import { afterEach, beforeEach, describe, expect, it, vi, type Mock } from 'vite
 import type { CanvasNode } from '../../../types';
 import { I18nProvider } from '../../../i18n';
 import { AppShellProvider } from '../../shell/AppShellProvider';
-import type { ChatDeliveryReceipt } from '../../chat/ChatTargetContext';
+import type { ChatDeliveryReceipt } from '../../../agent-chat/target';
 import { useCanvasNodeViewModel } from './useCanvasNodeViewModel';
 
 (globalThis as { IS_REACT_ACT_ENVIRONMENT?: boolean }).IS_REACT_ACT_ENVIRONMENT = true;
