+++ apps/canvas-workspace/src/renderer/src/components/canvas/CanvasNodeView/useCanvasNodeViewModel.ts
@@ -12,7 +12,7 @@ import type { CanvasNode } from '../../../types';
 import type { ResizeEdge } from '../../../hooks/useNodeResize';
 import { isImeComposing } from '../../../utils/ime';
 import { collectContainerDescendants } from '../../../utils/frameHierarchy';
-import { useChatDeliveryNotifier } from '../../chat/useChatDeliveryNotifier';
+import { useChatDeliveryNotifier } from '../../../agent-chat/target/useChatDeliveryNotifier';
 import { FULLSCREEN_NODE_TYPES } from './constants';
 import type { CanvasNodeViewProps } from './types';
 import {
