+++ apps/canvas-workspace/src/renderer/src/components/chat/ChatPageBody.tsx
@@ -14,9 +14,9 @@ import { useAppShell } from '../shell/AppShellProvider';
 import type { AgentNewSessionResult, AgentScope, WorkspaceOption } from './types';
 import { useI18n } from '../../i18n';
 import { isImeComposing } from '../../utils/ime';
-import type { ChatContextSnapshot, ChatExecutionPolicy, ChatTarget } from './ChatTargetContext';
+import type { ChatContextSnapshot, ChatExecutionPolicy, ChatTarget } from '../../agent-chat/target';
 import { chatScopeId } from './chatScope';
-import { useRegisterChatTarget } from './useRegisterChatTarget';
+import { useRegisterChatTarget } from '../../agent-chat/target/useRegisterChatTarget';
 import { ChatConversationStatus } from './ChatConversationStatus';
 import { useChatPageTargetContext } from './hooks/useChatPageTargetContext';
 import { useChatPageJumpNavigation } from './hooks/useChatPageJumpNavigation';
