--- apps/canvas-workspace/src/renderer/src/components/chat/ChatTargetContext.tsx
@@ -1,1 +1,0 @@
-export * from '../../agent-chat/target';
