+++ apps/canvas-workspace/src/renderer/src/components/chat/__tests__/ChatPage.scope-switch.test.tsx
@@ -2,7 +2,7 @@
 import { act } from 'react';
 import { createRoot, type Root } from 'react-dom/client';
 import { afterEach, describe, expect, it, vi } from 'vitest';
-import type { ChatTarget } from '../ChatTargetContext';
+import type { ChatTarget } from '../../../agent-chat/target';
 
 (globalThis as { IS_REACT_ACT_ENVIRONMENT?: boolean }).IS_REACT_ACT_ENVIRONMENT = true;
