--- apps/canvas-workspace/src/renderer/src/components/chat/useChatDeliveryNotifier.ts
@@ -1,4 +1,0 @@
-export {
-  useChatDeliveryNotifier,
-  type ChatDeliveryNotice,
-} from '../../agent-chat/target/useChatDeliveryNotifier';
