--- apps/canvas-workspace/src/renderer/src/components/chat/useRegisterChatTarget.ts
@@ -1,1 +1,0 @@
-export { useRegisterChatTarget } from '../../agent-chat/target/useRegisterChatTarget';
