--- apps/canvas-workspace/src/renderer/src/components/chat/utils/sessionScope.ts
@@ -1,4 +1,0 @@
-export {
-  createChatPageSessionTarget,
-  scopeFromSessionStoreId,
-} from '../../../agent-chat/target/sessionScope';
