+++ apps/canvas-workspace/src/renderer/src/components/dock/LinkDrawer/__tests__/dom-selection.test.tsx
@@ -4,7 +4,7 @@ import { createRoot, type Root } from 'react-dom/client';
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 import { AppShellProvider } from '../../../shell/AppShellProvider';
 import { I18nProvider } from '../../../../i18n';
-import { ChatTargetProvider } from '../../../chat/ChatTargetContext';
+import { ChatTargetProvider } from '../../../../agent-chat/target';
 import { LinkTabView } from '..';
 
 (globalThis as { IS_REACT_ACT_ENVIRONMENT?: boolean }).IS_REACT_ACT_ENVIRONMENT = true;
