+++ apps/canvas-workspace/src/renderer/src/components/dock/RightDock/NodeDetailDockTab.tsx
@@ -4,7 +4,7 @@ import { useKnowledgeTags, useWorkspaceNode, useWorkspaceNodeList } from '../../
 import { getNodeTitle } from '../../../views/WorkspaceNodes/utils';
 import { useI18n } from '../../../i18n';
 import type { AgentContextTabRef } from '../../../types';
-import type { ChatDeliveryReceipt } from '../../chat/ChatTargetContext';
+import type { ChatDeliveryReceipt } from '../../../agent-chat/target';
 import { TabChatAction } from './TabChatAction';
 
 interface NodeDetailDockTabProps {
