+++ apps/canvas-workspace/src/renderer/src/components/shell/Workbench/__tests__/ChatDockLifecycle.test.tsx
@@ -8,7 +8,7 @@ import {
   ChatTargetProvider,
   useActiveChatTarget,
   type ChatTarget,
-} from '../../../chat/ChatTargetContext';
+} from '../../../../agent-chat/target';
 import type { WorkbenchController } from '../useWorkbenchState';
 import { Workbench } from '../index';
 import { AppShellProvider } from '../../../shell/AppShellProvider';
@@ -39,7 +39,7 @@ vi.mock('../../../canvas/Canvas', () => ({
 
 vi.mock('../../../chat/lazy', async () => {
   const { useMemo: useReactMemo } = await import('react');
-  const { useRegisterChatTarget: useRegister } = await import('../../../chat/useRegisterChatTarget');
+  const { useRegisterChatTarget: useRegister } = await import('../../../../agent-chat/target/useRegisterChatTarget');
 
   return {
     ChatPanelLazy: ({
