+++ apps/canvas-workspace/src/renderer/src/components/shell/Workbench/__tests__/useChatInsertionBridge.test.tsx
@@ -3,7 +3,7 @@ import { act } from 'react';
 import { createRoot, type Root } from 'react-dom/client';
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 import type { AgentContextDomReviewComment, AgentContextDomSelectionRef, AgentContextTabRef, CanvasNode } from '../../../../types';
-import type { ChatTargetBroker } from '../../../chat/ChatTargetContext';
+import type { ChatTargetBroker } from '../../../../agent-chat/target';
 import { useChatInsertionBridge } from '../useChatInsertionBridge';
 
 (globalThis as { IS_REACT_ACT_ENVIRONMENT?: boolean }).IS_REACT_ACT_ENVIRONMENT = true;
